+++ packages/ReactDagEditor/src/components/Graph/GraphStateStore.tsx
@@ -10,26 +10,17 @@ import {
 } from "../../contexts";
 import { AlignmentLinesContext } from "../../contexts/AlignmentLinesContext";
 import { AutoZoomFitContext } from "../../contexts/AutoZoomFitContext";
-import {
-  GraphStateContext,
-  GraphValueContext,
-  IDispatch,
-  IDispatchCallback
-} from "../../contexts/GraphStateContext";
+import { GraphStateContext, GraphValueContext } from "../../contexts/GraphStateContext";
 import { GraphFeatures } from "../../Features";
+import { useGraphReducer } from "../../hooks/useGraphReducer";
 
 import { usePropsAPI } from "../../hooks/usePropsAPI";
 import { GraphCanvasEvent } from "../../models/event";
 import { ITransformMatrix } from "../../models/geometry";
 import { GraphModel } from "../../models/GraphModel";
-import { GraphBehavior } from "../../models/state";
 import { IPropsAPI } from "../../props-api/IPropsAPI";
-import { useGraphReducer } from "../../reducers/useGraphReducer";
-import { isViewportComplete, resetUndoStack } from "../../utils";
-import { batchedUpdates } from "../../utils/batchedUpdates";
+import { isViewportComplete } from "../../utils";
 import { graphController } from "../../utils/graphController";
-import { emptyDummyNodes } from "../dummyNodes";
-import { emptySelectBoxPosition } from "./SelectBox";
 
 export interface IGraphStateStoreProps<NodeData = unknown, EdgeData = unknown, PortData = unknown, Action = never> {
   /**
@@ -42,13 +33,12 @@ export interface IGraphStateStoreProps<NodeData = unknown, EdgeData = unknown, P
   data?: GraphModel<NodeData, EdgeData, PortData>;
   defaultTransformMatrix?: ITransformMatrix;
   middleware?: IGraphReducer<NodeData, EdgeData, PortData, Action>;
-  onStateChanged?: IDispatchCallback<NodeData, EdgeData, PortData>;
 }
 
 export function GraphStateStore<NodeData = unknown, EdgeData = unknown, PortData = unknown, Action = never>(
   props: React.PropsWithChildren<IGraphStateStoreProps<NodeData, EdgeData, PortData, Action>>
 ): React.ReactElement {
-  const { defaultTransformMatrix = EMPTY_TRANSFORM_MATRIX, middleware, onStateChanged } = props;
+  const { defaultTransformMatrix = EMPTY_TRANSFORM_MATRIX, middleware } = props;
 
   const propsAPI = usePropsAPI<NodeData, EdgeData, PortData>();
   React.useImperativeHandle(props.propsAPIRef, () => propsAPI, [propsAPI]);
@@ -62,56 +52,15 @@ export function GraphStateStore<NodeData = unknown, EdgeData = unknown, PortData
     }),
     [enabledFeatures, graphConfig]
   );
-  const reducer = useGraphReducer(reducerContext, middleware as IGraphReducer);
 
-  const [state, dispatchImpl] = React.useReducer(reducer, undefined, () => ({
-    data: resetUndoStack(props.data ?? GraphModel.empty()),
-    viewport: {
-      rect: undefined,
+  const [state, dispatch] = useGraphReducer(
+    {
+      data: props.data,
       transformMatrix: defaultTransformMatrix
     },
-    behavior: GraphBehavior.default,
-    dummyNodes: emptyDummyNodes(),
-    alignmentLines: [],
-    activeKeys: new Set<string>(),
-    selectBoxPosition: emptySelectBoxPosition(),
-    connectState: undefined
-  }));
-
-  const sideEffects = React.useMemo<IDispatchCallback[]>(() => [], []);
-  const prevStateRef = React.useRef(state);
-
-  const dispatch: IDispatch = React.useCallback(
-    (action, callback) => {
-      if (callback) {
-        sideEffects.push(callback);
-      }
-      dispatchImpl(action);
-    },
-    [sideEffects]
+    reducerContext,
+    middleware
   );
-
-  React.useEffect((): void => {
-    const prevState = prevStateRef.current;
-    if (prevState === state) {
-      return;
-    }
-    prevStateRef.current = state;
-    if (onStateChanged) {
-      sideEffects.unshift(onStateChanged);
-    }
-    batchedUpdates(() => {
-      sideEffects.forEach(callback => {
-        try {
-          callback(state, prevState);
-        } catch (e) {
-          // eslint-disable-next-line no-console
-          console.error(e);
-        }
-      });
-      sideEffects.length = 0;
-    });
-  });
 
   const contextValue = React.useMemo(
     () => ({
